💥🤯 Great refactoring, many tests, IO task rework, etc.
Affected tickets:
* Resolves #7: current task is now passed into the block (for both
  normal and IO tasks), with tests.
* Resolves #12: using `run` in sink context now explodes properly
  (tested for both normal and IO tasks).
* Resolves #13: no more shell injection through filenames, tested with
  extra tests for filenames starting with `--`.
* Resolves #16: there are now helper functions to test sakefiles
  without the need to `.execute` tasks from within the test file.  It
  works by creating a temp directory with a given Sakefile and running
  `sake` as external command (stdout, stderr, exit code and signals
  are checked).
* Resolves #18: you can now pass an IO object instead of a Str and it
  will automatically dispatch to the right sub.
* There is some groundwork for #17, and maybe it already works. I
  can't tell if it does because there are no tests for it yet.
* Also some minimal groundwork for #9 (parallel execution).
* Additionally, issue #14 should be more approachable now (because
  deps are now resolved).
* Issue #15 (‘default’ task) was reworked a bit (with no notable
  functional changes).

This commit splits Sake.pm6 into separate files for convenience. The
hierarchy is perhaps not entirely right, but now it is much easier to
refactor it further with all the tests.

Because almost all lines were touched, this commit introduces some
major code style changes (to my personal preference, e.g. unicode
quotes). I'm not insisting on that style, so if anybody cares enough
it should be possible to submit a PR with unicode stuff autoreplaced.

I wish this commit was split into multiple digestable commits, but it
was a single refactoring effort that I ended up shelving for a few
months anyway, so I'm happy that it goes in at all.